refactor(interpreter): split comprehensions and subscripts
