+++ src/interpreter/src/expressions/mod.rs
@@ -7,15 +7,38 @@ use std::collections::HashMap;
 use std::collections::HashSet;
 
 mod environment;
+mod comprehensions;
 mod ranges;
 mod registration;
+mod subscripts;
 mod variables;
 
+#[cfg(feature = "matrix_comprehensions")]
+pub use comprehensions::{
+  MatrixComprehensionDefine, ValueMatrixComprehension, matrix_comprehension,
+};
+#[cfg(feature = "set_comprehensions")]
+pub use comprehensions::{SetComprehensionDefine, ValueSetComprehension, set_comprehension};
 pub(crate) use environment::DeferredExpressionSolveScope;
 use environment::expression_solves_deferred;
 pub use ranges::range;
 use registration::{
   register_expression_function_batch, register_initialized_expression_function,
+};
+#[cfg(all(feature = "subscript", feature = "access"))]
+pub use subscripts::subscript;
+#[cfg(feature = "subscript_formula")]
+pub use subscripts::{subscript_formula, subscript_formula_ix};
+#[cfg(feature = "subscript_range")]
+pub use subscripts::subscript_range;
+#[cfg(all(feature = "subscript_slice", feature = "access"))]
+pub use subscripts::slice;
+#[cfg(feature = "subscript_formula")]
+pub(crate) use subscripts::{
+  current_string_access_expression_live, mark_current_string_access_expression_live,
+  mark_string_access_value_live, reset_current_string_access_expression_live,
+  string_access_input_is_live, string_access_value_is_marked_live,
+  take_current_string_access_expression_live,
 };
 use variables::{addressed_identifier_hash, addressed_identifier_name};
 pub use variables::var;
@@ -52,981 +75,6 @@ pub fn expression(expr: &Expression, env: Option<&Environment>, p: &InterpreterE
         x => Err(MechError::new(FeatureNotEnabledError, None)
             .with_compiler_loc()
             .with_tokens(x.tokens())),
-    }
-}
-
-#[cfg(any(feature = "set_comprehensions", feature = "matrix_comprehensions"))]
-fn comprehension_environments(
-    qualifiers: &[ComprehensionQualifier],
-    comprehension_id: u64,
-    p: &InterpreterExecution<'_>,
-) -> MResult<(Vec<Environment>, Interpreter)> {
-    let mut envs: Vec<Environment> = vec![HashMap::new()];
-    let mut new_p: Interpreter = (**p).clone();
-    new_p.id = comprehension_id;
-    new_p.clear_plan();
-    for qual in qualifiers {
-        envs = match qual {
-            ComprehensionQualifier::Generator((pttrn, expr)) => {
-                let compiled = crate::patterns::compile_pattern(pttrn, None, &new_p)?;
-                let mut new_envs = Vec::new();
-                for env in &envs {
-                    let collection = p.with_interpreter(
-                        &new_p,
-                        |execution| expression(
-                            expr,
-                            Some(env),
-                            execution,
-                        ),
-                    )?;
-                    for elmnt in comprehension_generator_values(&collection)? {
-                        let mut new_env = env.clone();
-                        let pattern_match = p.with_interpreter(
-                            &new_p,
-                            |execution| {
-                                crate::patterns::match_compiled_pattern_with_environment_constraints(
-                                    &compiled,
-                                    &elmnt,
-                                    &new_env,
-                                    execution,
-                                )
-                            },
-                        )?;
-                        if pattern_match.matched {
-                            crate::patterns::EnvironmentBindingSink::new(&mut new_env)
-                                .commit(&pattern_match)?;
-                            new_envs.push(new_env);
-                        }
-                    }
-                }
-                new_envs
-            }
-            ComprehensionQualifier::Filter(expr) => envs
-                .into_iter()
-                .filter(|env| {
-                    let result = p.with_interpreter(
-                        &new_p,
-                        |execution| expression(
-                            expr,
-                            Some(env),
-                            execution,
-                        ),
-                    );
-                    match result {
-                        Ok(Value::Bool(v)) => v.borrow().clone(),
-                        Ok(_) => false,
-                        Err(_) => false,
-                    }
-                })
-                .collect(),
-            ComprehensionQualifier::Let(var_def) => envs
-                .into_iter()
-                .map(|mut env| -> MResult<_> {
-                    let val = p.with_interpreter(
-                        &new_p,
-                        |execution| expression(
-                            &var_def.expression,
-                            Some(&env),
-                            execution,
-                        ),
-                    )?;
-                    env.insert(var_def.var.name.hash(), val);
-                    Ok(env)
-                })
-                .collect::<MResult<Vec<_>>>()?,
-        };
-    }
-    Ok((envs, new_p))
-}
-
-#[cfg(any(feature = "set_comprehensions", feature = "matrix_comprehensions"))]
-fn comprehension_generator_values(collection: &Value) -> MResult<Vec<Value>> {
-    match collection {
-        #[cfg(feature = "set")]
-        Value::Set(mset) => Ok(mset.borrow().set.iter().cloned().collect()),
-        #[cfg(feature = "matrix")]
-        Value::MatrixIndex(matrix) => Ok(matrix
-            .as_vec()
-            .into_iter()
-            .map(|value| Value::Index(Ref::new(value)))
-            .collect()),
-        #[cfg(all(feature = "matrix", feature = "bool"))]
-        Value::MatrixBool(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "u8"))]
-        Value::MatrixU8(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "u16"))]
-        Value::MatrixU16(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "u32"))]
-        Value::MatrixU32(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "u64"))]
-        Value::MatrixU64(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "u128"))]
-        Value::MatrixU128(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "i8"))]
-        Value::MatrixI8(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "i16"))]
-        Value::MatrixI16(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "i32"))]
-        Value::MatrixI32(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "i64"))]
-        Value::MatrixI64(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "i128"))]
-        Value::MatrixI128(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "f32"))]
-        Value::MatrixF32(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "f64"))]
-        Value::MatrixF64(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "string"))]
-        Value::MatrixString(matrix) => Ok(matrix.as_vec().into_iter().map(Value::from).collect()),
-        #[cfg(all(feature = "matrix", feature = "rational"))]
-        Value::MatrixR64(matrix) => Ok(matrix
-            .as_vec()
-            .into_iter()
-            .map(|value| value.to_value())
-            .collect()),
-        #[cfg(all(feature = "matrix", feature = "complex"))]
-        Value::MatrixC64(matrix) => Ok(matrix
-            .as_vec()
-            .into_iter()
-            .map(|value| value.to_value())
-            .collect()),
-        #[cfg(feature = "matrix")]
-        Value::MatrixValue(matrix) => Ok(matrix.as_vec()),
-        Value::MutableReference(reference) => comprehension_generator_values(&reference.borrow()),
-        x => Err(
-            MechError::new(ComprehensionGeneratorError { found: x.kind() }, None)
-                .with_compiler_loc(),
-        ),
-    }
-}
-
-#[cfg(any(feature = "set_comprehensions", feature = "matrix_comprehensions"))]
-fn detach_comprehension_value(value: &Value) -> Value {
-    match value {
-        Value::MutableReference(reference) => reference.borrow().clone(),
-        _ => value.clone(),
-    }
-}
-
-#[cfg(feature = "set_comprehensions")]
-#[derive(Debug)]
-pub struct ValueSetComprehension {
-    pub arguments: Vec<Value>,
-    pub out: Ref<MechSet>,
-}
-#[cfg(all(feature = "set_comprehensions", feature = "functions"))]
-impl MechFunctionImpl for ValueSetComprehension {
-    fn solve(&self) {
-        let args = self
-            .arguments
-            .iter()
-            .map(detach_comprehension_value)
-            .collect::<Vec<Value>>();
-        *self.out.borrow_mut() = MechSet::from_vec(args);
-    }
-    fn out(&self) -> Value {
-        Value::Set(self.out.clone())
-    }
-    fn to_string(&self) -> String {
-        format!("{:#?}", self)
-    }
-
-  fn transaction_state_values(&self) -> MResult<Vec<Value>> {
-    Ok(self.reactive_output_values())
-  }
-}
-#[cfg(all(feature = "set_comprehensions", feature = "functions"))]
-impl MechFunctionFactory for ValueSetComprehension {
-    fn new(args: FunctionArgs) -> MResult<Box<dyn MechFunction>> {
-        match args {
-            FunctionArgs::Nullary(Value::Set(out)) => Ok(Box::new(ValueSetComprehension {
-                arguments: Vec::new(),
-                out,
-            })),
-            FunctionArgs::Nullary(out) => Err(MechError::new(
-                SetComprehensionOutputKindMismatchError { found: out.kind() },
-                None,
-            )
-            .with_compiler_loc()),
-            _ => Err(MechError::new(
-                IncorrectNumberOfArguments {
-                    expected: 0,
-                    found: args.len(),
-                },
-                None,
-            )
-            .with_compiler_loc()),
-        }
-    }
-}
-#[cfg(all(feature = "set_comprehensions", feature = "compiler"))]
-impl MechFunctionCompiler for ValueSetComprehension {
-    fn compile(&self, ctx: &mut CompileCtx) -> MResult<Register> {
-        compile_nullop!(
-            "set/comprehension",
-            self.out,
-            ctx,
-            FeatureFlag::Builtin(FeatureKind::SetComprehensions)
-        );
-    }
-}
-#[cfg(all(feature = "set_comprehensions", feature = "functions"))]
-register_descriptor! {
-  FunctionDescriptor {
-    name: "set/comprehension",
-    ptr: ValueSetComprehension::new,
-  }
-}
-#[cfg(feature = "set_comprehensions")]
-pub struct SetComprehensionDefine {}
-#[cfg(all(feature = "set_comprehensions", feature = "functions"))]
-impl NativeFunctionCompiler for SetComprehensionDefine {
-    fn compile(&self, arguments: &Vec<Value>) -> MResult<Box<dyn MechFunction>> {
-        Ok(Box::new(ValueSetComprehension {
-            arguments: arguments.clone(),
-            out: Ref::new(MechSet::from_vec(arguments.clone())),
-        }))
-    }
-}
-#[cfg(all(feature = "set_comprehensions", feature = "functions"))]
-register_descriptor! {
-  FunctionCompilerDescriptor {
-    name: "set/comprehension",
-    ptr: &SetComprehensionDefine{},
-  }
-}
-
-#[cfg(feature = "matrix_comprehensions")]
-#[derive(Debug)]
-pub struct ValueMatrixComprehension {
-    pub arguments: Vec<Value>,
-    pub out: Ref<Value>,
-}
-#[cfg(all(feature = "matrix_comprehensions", feature = "functions"))]
-impl MechFunctionImpl for ValueMatrixComprehension {
-    fn solve(&self) {
-        let args = self
-            .arguments
-            .iter()
-            .map(detach_comprehension_value)
-            .collect::<Vec<Value>>();
-        let out = if args.is_empty() {
-            Value::MatrixValue(Matrix::from_vec(vec![], 0, 0))
-        } else {
-            let fxn = MatrixHorzCat {}
-                .compile(&args)
-                .expect("matrix/comprehension input kinds changed to incompatible values");
-            fxn.solve();
-            fxn.out()
-        };
-        *self.out.borrow_mut() = out;
-    }
-    fn out(&self) -> Value {
-        self.out.borrow().clone()
-    }
-    fn transaction_state_values(&self) -> MResult<Vec<Value>> {
-        Ok(vec![Value::MutableReference(self.out.clone())])
-    }
-    fn to_string(&self) -> String {
-        format!("{:#?}", self)
-    }
-}
-
-#[cfg(all(feature = "matrix_comprehensions", feature = "functions"))]
-impl MechFunctionFactory for ValueMatrixComprehension {
-    fn new(args: FunctionArgs) -> MResult<Box<dyn MechFunction>> {
-        match args {
-            FunctionArgs::Nullary(out) => Ok(Box::new(ValueMatrixComprehension {
-                arguments: Vec::new(),
-                out: Ref::new(out),
-            })),
-            _ => Err(MechError::new(
-                IncorrectNumberOfArguments {
-                    expected: 0,
-                    found: args.len(),
-                },
-                None,
-            )
-            .with_compiler_loc()),
-        }
-    }
-}
-#[cfg(all(feature = "matrix_comprehensions", feature = "compiler"))]
-impl MechFunctionCompiler for ValueMatrixComprehension {
-    fn compile(&self, ctx: &mut CompileCtx) -> MResult<Register> {
-        compile_nullop!(
-            "matrix/comprehension",
-            self.out,
-            ctx,
-            FeatureFlag::Builtin(FeatureKind::MatrixComprehensions)
-        );
-    }
-}
-#[cfg(all(feature = "matrix_comprehensions", feature = "functions"))]
-register_descriptor! {
-  FunctionDescriptor {
-    name: "matrix/comprehension",
-    ptr: ValueMatrixComprehension::new,
-  }
-}
-#[cfg(feature = "matrix_comprehensions")]
-pub struct MatrixComprehensionDefine {}
-#[cfg(all(feature = "matrix_comprehensions", feature = "functions"))]
-impl NativeFunctionCompiler for MatrixComprehensionDefine {
-    fn compile(&self, arguments: &Vec<Value>) -> MResult<Box<dyn MechFunction>> {
-        let out = if arguments.is_empty() {
-            Value::MatrixValue(Matrix::from_vec(vec![], 0, 0))
-        } else {
-            let fxn = MatrixHorzCat {}.compile(arguments)?;
-            fxn.solve();
-            fxn.out()
-        };
-        Ok(Box::new(ValueMatrixComprehension {
-            arguments: arguments.clone(),
-            out: Ref::new(out),
-        }))
-    }
-}
-#[cfg(all(feature = "matrix_comprehensions", feature = "functions"))]
-register_descriptor! {
-  FunctionCompilerDescriptor {
-    name: "matrix/comprehension",
-    ptr: &MatrixComprehensionDefine{},
-  }
-}
-
-#[cfg(feature = "set_comprehensions")]
-pub fn set_comprehension(set_comp: &SetComprehension, p: &InterpreterExecution<'_>) -> MResult<Value> {
-    let comprehension_id = hash_str(&format!("{:?}", set_comp));
-    let (envs, new_p) = comprehension_environments(&set_comp.qualifiers, comprehension_id, p)?;
-    let mut values = Vec::new();
-    for env in envs {
-        let val = p.with_interpreter(
-            &new_p,
-            |execution| expression(
-                &set_comp.expression,
-                Some(&env),
-                execution,
-            ),
-        )?;
-        values.push(val);
-    }
-    let functions = p.functions();
-    let set_define_id = hash_str("set/comprehension");
-    let set_define = {
-        functions
-            .borrow()
-            .function_compilers
-            .get(&set_define_id)
-            .cloned()
-    };
-    match set_define {
-        Some(compiler) => execute_native_function_compiler(compiler, &values, p),
-        None => Err(MechError::new(
-            MissingFunctionError {
-                function_id: set_define_id,
-            },
-            None,
-        )
-        .with_compiler_loc()),
-    }
-}
-
-#[cfg(feature = "matrix_comprehensions")]
-pub fn matrix_comprehension(matrix_comp: &MatrixComprehension, p: &InterpreterExecution<'_>) -> MResult<Value> {
-    let comprehension_id = hash_str(&format!("{:?}", matrix_comp));
-    let (envs, new_p) = comprehension_environments(&matrix_comp.qualifiers, comprehension_id, p)?;
-    let mut values = Vec::new();
-    for env in envs {
-        values.push(p.with_interpreter(
-            &new_p,
-            |execution| expression(
-                &matrix_comp.expression,
-                Some(&env),
-                execution,
-            ),
-        )?);
-    }
-    let functions = p.functions();
-    let horzcat_id = hash_str("matrix/comprehension");
-    let horzcat = {
-        functions
-            .borrow()
-            .function_compilers
-            .get(&horzcat_id)
-            .cloned()
-    };
-    match horzcat {
-        Some(compiler) => execute_native_function_compiler(compiler, &values, p),
-        None => Err(MechError::new(
-            MissingFunctionError {
-                function_id: horzcat_id,
-            },
-            None,
-        )
-        .with_compiler_loc()),
-    }
-}
-
-#[cfg(all(feature = "subscript_slice", feature = "access"))]
-pub fn slice(slc: &Slice, env: Option<&Environment>, p: &InterpreterExecution<'_>) -> MResult<Value> {
-    let id = addressed_identifier_hash(&slc.name, &slc.context);
-    let name = addressed_identifier_name(&slc.name, &slc.context);
-    let val: Value = if let Some(env) = env {
-        if let Some(val) = env.get(&id) {
-            val.clone()
-        } else {
-            // fallback to global symbols
-            {
-                let symbols = p.symbols();
-                let symbols_brrw = symbols.borrow();
-                match symbols_brrw.get(id) {
-                    Some(val) => match symbols_brrw.get_mutable(id) {
-                        Some(_) => Value::MutableReference(val.clone()),
-                        None => val.borrow().clone(),
-                    },
-                    None => {
-                        return Err(MechError::new(UndefinedVariableError { id, name: name.clone() }, None)
-                            .with_compiler_loc()
-                            .with_tokens(slc.tokens()));
-                    }
-                }
-            }
-        }
-    } else {
-        let symbols = p.symbols();
-        let symbols_brrw = symbols.borrow();
-        match symbols_brrw.get(id) {
-            Some(val) => match symbols_brrw.get_mutable(id) {
-                Some(_) => Value::MutableReference(val.clone()),
-                None => val.borrow().clone(),
-            },
-            None => {
-                return Err(MechError::new(UndefinedVariableError { id, name: name.clone() }, None)
-                    .with_compiler_loc()
-                    .with_tokens(slc.tokens()));
-            }
-        }
-    };
-    let mut v = val;
-    for s in &slc.subscript {
-        v = subscript(s, &v, env, p)?;
-    }
-    Ok(v)
-}
-
-#[cfg(feature = "subscript_formula")]
-pub fn subscript_formula(
-    sbscrpt: &Subscript,
-    env: Option<&Environment>,
-    p: &InterpreterExecution<'_>,
-) -> MResult<Value> {
-    match sbscrpt {
-        Subscript::Formula(fctr) => factor(fctr, env, p),
-        _ => unreachable!(),
-    }
-}
-
-#[cfg(feature = "subscript_formula")]
-pub fn subscript_formula_ix(
-    sbscrpt: &Subscript,
-    env: Option<&Environment>,
-    p: &InterpreterExecution<'_>,
-) -> MResult<Value> {
-    match sbscrpt {
-        Subscript::Formula(fctr) => {
-            let result = factor(fctr, env, p)?;
-            result.as_index()
-        }
-        _ => unreachable!(),
-    }
-}
-
-
-#[cfg(feature = "subscript_formula")]
-pub(crate) fn reset_current_string_access_expression_live(p: &InterpreterExecution<'_>) {
-    *p.current_string_access_expression_live.borrow_mut() = false;
-}
-
-#[cfg(feature = "subscript_formula")]
-pub(crate) fn current_string_access_expression_live(p: &InterpreterExecution<'_>) -> bool {
-    *p.current_string_access_expression_live.borrow()
-}
-
-#[cfg(feature = "subscript_formula")]
-pub(crate) fn take_current_string_access_expression_live(p: &InterpreterExecution<'_>) -> bool {
-    let value = *p.current_string_access_expression_live.borrow();
-    *p.current_string_access_expression_live.borrow_mut() = false;
-    value
-}
-
-#[cfg(feature = "subscript_formula")]
-pub(crate) fn mark_current_string_access_expression_live(p: &InterpreterExecution<'_>) {
-    *p.current_string_access_expression_live.borrow_mut() = true;
-}
-
-#[cfg(feature = "subscript_formula")]
-fn string_access_scalar_addr(value: &Value) -> Option<usize> {
-    match value {
-        Value::MutableReference(reference) => string_access_scalar_addr(&reference.borrow()),
-        Value::Typed(value, _) => string_access_scalar_addr(value),
-        Value::String(value) => Some(value.addr()),
-        Value::Index(value) => Some(value.addr()),
-
-        #[cfg(feature = "u8")]
-        Value::U8(value) => Some(value.addr()),
-        #[cfg(feature = "u16")]
-        Value::U16(value) => Some(value.addr()),
-        #[cfg(feature = "u32")]
-        Value::U32(value) => Some(value.addr()),
-        #[cfg(feature = "u64")]
-        Value::U64(value) => Some(value.addr()),
-        #[cfg(feature = "u128")]
-        Value::U128(value) => Some(value.addr()),
-
-        #[cfg(feature = "i8")]
-        Value::I8(value) => Some(value.addr()),
-        #[cfg(feature = "i16")]
-        Value::I16(value) => Some(value.addr()),
-        #[cfg(feature = "i32")]
-        Value::I32(value) => Some(value.addr()),
-        #[cfg(feature = "i64")]
-        Value::I64(value) => Some(value.addr()),
-        #[cfg(feature = "i128")]
-        Value::I128(value) => Some(value.addr()),
-
-        #[cfg(feature = "f32")]
-        Value::F32(value) => Some(value.addr()),
-        #[cfg(feature = "f64")]
-        Value::F64(value) => Some(value.addr()),
-
-        _ => None,
-    }
-}
-
-#[cfg(feature = "subscript_formula")]
-pub(crate) fn mark_string_access_value_live(p: &InterpreterExecution<'_>, value: &Value) {
-    if let Some(addr) = string_access_scalar_addr(value) {
-        p.string_access_live_values.borrow_mut().insert(addr);
-    }
-}
-
-#[cfg(feature = "subscript_formula")]
-pub(crate) fn string_access_value_is_marked_live(p: &InterpreterExecution<'_>, value: &Value) -> bool {
-    string_access_scalar_addr(value)
-        .map(|addr| p.string_access_live_values.borrow().contains(&addr))
-        .unwrap_or(false)
-}
-
-#[cfg(feature = "subscript_formula")]
-fn subscript_formula_is_mutable_symbol(
-    sbscrpt: &Subscript,
-    env: Option<&Environment>,
-    p: &InterpreterExecution<'_>,
-) -> bool {
-    if env.is_some() {
-        return false;
-    }
-    let Subscript::Formula(fctr) = sbscrpt else {
-        return false;
-    };
-    let Factor::Expression(expr) = fctr else {
-        return false;
-    };
-    let Expression::Var(var) = expr.as_ref() else {
-        return false;
-    };
-    let id = addressed_identifier_hash(&var.name, &var.context);
-    let state_brrw = p.state.borrow();
-    let symbols_brrw = state_brrw.symbol_table.borrow();
-    symbols_brrw.get_mutable(id).is_some()
-}
-
-#[cfg(feature = "subscript_formula")]
-fn mutable_reference_is_mutable_symbol(reference: &MutableReference, p: &InterpreterExecution<'_>) -> bool {
-    let state_brrw = p.state.borrow();
-    let symbols_brrw = state_brrw.symbol_table.borrow();
-    symbols_brrw
-        .mutable_variables
-        .values()
-        .any(|symbol| symbol.same_handle(reference))
-}
-
-#[cfg(feature = "subscript_formula")]
-fn value_is_mutable_symbol_reference(value: &Value, p: &InterpreterExecution<'_>) -> bool {
-    match value {
-        Value::MutableReference(reference) => mutable_reference_is_mutable_symbol(reference, p),
-        _ => false,
-    }
-}
-
-#[cfg(feature = "subscript_formula")]
-fn mutable_reference_is_live_plan_output(reference: &MutableReference, p: &InterpreterExecution<'_>) -> bool {
-    let current = reference.borrow();
-    string_access_value_is_marked_live(p, &current)
-}
-
-#[cfg(feature = "subscript_formula")]
-fn string_access_argument_is_live(value: &Value, p: &InterpreterExecution<'_>) -> bool {
-    string_access_value_is_marked_live(p, value)
-}
-
-#[cfg(feature = "subscript_formula")]
-pub(crate) fn string_access_input_is_live(value: &Value, p: &InterpreterExecution<'_>) -> bool {
-    value_is_mutable_symbol_reference(value, p) || string_access_argument_is_live(value, p)
-}
-
-#[cfg(feature = "subscript_formula")]
-fn string_access_source_argument(value: &Value, p: &InterpreterExecution<'_>) -> Value {
-    match value {
-        Value::MutableReference(reference)
-            if matches!(value.deref_kind(), ValueKind::String)
-                && !mutable_reference_is_mutable_symbol(reference, p)
-                && !mutable_reference_is_live_plan_output(reference, p) =>
-        {
-            reference.borrow().clone()
-        }
-        _ => value.clone(),
-    }
-}
-
-#[cfg(feature = "subscript_formula")]
-fn string_access_index_argument(
-    raw_index: Value,
-    sbscrpt: &Subscript,
-    env: Option<&Environment>,
-    p: &InterpreterExecution<'_>,
-) -> MResult<Value> {
-    match &raw_index {
-        Value::MutableReference(reference)
-            if subscript_formula_is_mutable_symbol(sbscrpt, env, p)
-                || mutable_reference_is_live_plan_output(reference, p) =>
-        {
-            reference.borrow().as_index()?;
-            Ok(raw_index)
-        }
-        _ => raw_index.as_index(),
-    }
-}
-
-#[cfg(feature = "subscript_range")]
-pub fn subscript_range(
-    sbscrpt: &Subscript,
-    env: Option<&Environment>,
-    p: &InterpreterExecution<'_>,
-) -> MResult<Value> {
-    match sbscrpt {
-        Subscript::Range(rng) => {
-            let result = range(rng, env, p)?;
-            match result.as_vecusize() {
-                Ok(v) => Ok(v.to_value()),
-                Err(_) => Err(MechError::new(
-                    InvalidIndexKindError {
-                        kind: result.kind(),
-                    },
-                    None,
-                )
-                .with_compiler_loc()
-                .with_tokens(rng.tokens())),
-            }
-        }
-        _ => unreachable!(),
-    }
-}
-
-#[cfg(all(feature = "subscript", feature = "access"))]
-pub fn subscript(
-    sbscrpt: &Subscript,
-    val: &Value,
-    env: Option<&Environment>,
-    p: &InterpreterExecution<'_>,
-) -> MResult<Value> {
-    let plan = p.plan();
-    match sbscrpt {
-        #[cfg(feature = "table")]
-        Subscript::Dot(x) => {
-            let key = x.hash();
-            let fxn_input: Vec<Value> = vec![val.clone(), Value::Id(key)];
-            #[cfg(feature = "record")]
-            if matches!(val.deref_kind(), ValueKind::Record(..)) {
-                let function = AccessColumn {}.compile(&fxn_input)?;
-                return register_initialized_expression_function(&plan, function, &[]);
-            }
-            let new_fxn = AccessColumn {}.compile(&fxn_input)?;
-            if !expression_solves_deferred(p) {
-              new_fxn.solve();
-            }
-            let res = new_fxn.out();
-            plan.borrow_mut().push(new_fxn);
-            return Ok(res);
-        }
-        Subscript::DotInt(x) => {
-            let mut fxn_input = vec![val.clone()];
-            let result = real(&x.clone(), p)?;
-            fxn_input.push(result.as_index()?);
-            match val.deref_kind() {
-                #[cfg(feature = "matrix")]
-                ValueKind::Matrix(..) => {
-                    let new_fxn = MatrixAccessScalar {}.compile(&fxn_input)?;
-                    if !expression_solves_deferred(p) {
-                      new_fxn.solve();
-                    }
-                    let res = new_fxn.out();
-                    plan.borrow_mut().push(new_fxn);
-                    return Ok(res);
-                }
-                #[cfg(feature = "tuple")]
-                ValueKind::Tuple(..) => {
-                    let function = TupleAccess {}.compile(&fxn_input)?;
-                    return register_initialized_expression_function(&plan, function, &[]);
-                }
-                /*ValueKind::Record(_) => {
-                  let new_fxn = RecordAccessScalar{}.compile(&fxn_input)?;
-                  new_fxn.solve();
-                  let res = new_fxn.out();
-                  plan.borrow_mut().push(new_fxn);
-                  return Ok(res);
-                },*/
-                _ => todo!("Implement access for dot int"),
-            }
-        }
-        #[cfg(feature = "swizzle")]
-        Subscript::Swizzle(x) => {
-            let mut keys = x
-                .iter()
-                .map(|x| Value::Id(x.hash()))
-                .collect::<Vec<Value>>();
-            let mut fxn_input: Vec<Value> = vec![val.clone()];
-            fxn_input.append(&mut keys);
-            let new_fxn = AccessSwizzle {}.compile(&fxn_input)?;
-            if !expression_solves_deferred(p) {
-              new_fxn.solve();
-            }
-            let res = new_fxn.out();
-            plan.borrow_mut().push(new_fxn);
-            return Ok(res);
-        }
-        Subscript::Brace(subs) => {
-            let mut fxn_input = vec![val.clone()];
-            match &subs[..] {
-                #[cfg(feature = "subscript_formula")]
-                [Subscript::Formula(ix)] => {
-                    let result = subscript_formula(&subs[0], env, p)?;
-                    let shape = result.shape();
-                    fxn_input.push(result);
-                    match shape[..] {
-                        [1, 1] => { plan.borrow_mut().push(AccessScalar {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "subscript_range")]
-                        [n, 1] => { plan.borrow_mut().push(AccessRange {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "subscript_range")]
-                        [1, n] => { plan.borrow_mut().push(AccessRange {}.compile(&fxn_input)?); }
-                        _ => todo!(),
-                    }
-                }
-                #[cfg(feature = "subscript_range")]
-                [Subscript::Range(ix)] => {
-                    let result = subscript_range(&subs[0], env, p)?;
-                    fxn_input.push(result);
-                    plan.borrow_mut().push(AccessRange {}.compile(&fxn_input)?);
-                }
-                /*[Subscript::All] => {
-                  fxn_input.push(Value::IndexAll);
-                  #[cfg(feature = "matrix")]
-                  plan.borrow_mut().push(MapAccessAll{}.compile(&fxn_input)?);
-                },*/
-                _ => {
-                    todo!("Implement brace subscript")
-                }
-            }
-            let plan_brrw = plan.borrow();
-            let mut new_fxn = &plan_brrw.last().unwrap();
-            if !expression_solves_deferred(p) {
-              new_fxn.solve();
-            }
-            let res = new_fxn.out();
-            return Ok(res);
-        }
-        #[cfg(feature = "subscript_slice")]
-        Subscript::Bracket(subs) => {
-            let string_source_is_live = matches!(val.deref_kind(), ValueKind::String)
-                && string_access_argument_is_live(val, p);
-            let mut fxn_input = if matches!(val.deref_kind(), ValueKind::String) {
-                vec![string_access_source_argument(val, p)]
-            } else {
-                vec![val.clone()]
-            };
-            match &subs[..] {
-                #[cfg(feature = "subscript_formula")]
-                [Subscript::Formula(ix)] => {
-                    let raw_index = subscript_formula(&subs[0], env, p)?;
-                    let index_arg = if matches!(val.deref_kind(), ValueKind::String) {
-                        string_access_index_argument(raw_index, &subs[0], env, p)?
-                    } else {
-                        raw_index.as_index()?
-                    };
-                    if matches!(val.deref_kind(), ValueKind::String)
-                        && matches!(fxn_input.first(), Some(Value::String(_)))
-                        && matches!(&index_arg, Value::Index(_))
-                    {
-                        let mode = if current_string_access_expression_live(p)
-                            || string_source_is_live
-                            || string_access_argument_is_live(&index_arg, p)
-                        {
-                            StringAccessCompileMode::LiveDirect
-                        } else {
-                            StringAccessCompileMode::Constant
-                        };
-                        set_next_string_access_compile_mode(mode);
-                    }
-                    let shape = index_arg.shape();
-                    fxn_input.push(index_arg);
-                    match shape[..] {
-                        [1, 1] => { plan.borrow_mut().push(AccessScalar {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "subscript_range")]
-                        [1, n] => { plan.borrow_mut().push(AccessRange {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "subscript_range")]
-                        [n, 1] => { plan.borrow_mut().push(AccessRange {}.compile(&fxn_input)?); }
-                        _ => todo!(),
-                    }
-                }
-                #[cfg(feature = "subscript_range")]
-                [Subscript::Range(ix)] => {
-                    let result = subscript_range(&subs[0], env, p)?;
-                    fxn_input.push(result);
-                    plan.borrow_mut().push(AccessRange {}.compile(&fxn_input)?);
-                }
-                [Subscript::All] => {
-                    fxn_input.push(Value::IndexAll);
-                    #[cfg(feature = "matrix")]
-                    plan.borrow_mut()
-                        .push(MatrixAccessAll {}.compile(&fxn_input)?);
-                }
-                [Subscript::All, Subscript::All] => todo!(),
-                #[cfg(feature = "subscript_formula")]
-                [Subscript::Formula(ix1), Subscript::Formula(ix2)] => {
-                    let result = subscript_formula_ix(&subs[0], env, p)?;
-                    let shape1 = result.shape();
-                    fxn_input.push(result);
-                    let result = subscript_formula_ix(&subs[1], env, p)?;
-                    let shape2 = result.shape();
-                    fxn_input.push(result);
-                    match ((shape1[0], shape1[1]), (shape2[0], shape2[1])) {
-                        #[cfg(feature = "matrix")]
-                        ((1, 1), (1, 1)) => { plan.borrow_mut().push(MatrixAccessScalarScalar {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "matrix")]
-                        ((1, 1), (m, 1)) => { plan.borrow_mut().push(MatrixAccessScalarRange {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "matrix")]
-                        ((n, 1), (1, 1)) => { plan.borrow_mut().push(MatrixAccessRangeScalar {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "matrix")]
-                        ((n, 1), (m, 1)) => { plan.borrow_mut().push(MatrixAccessRangeRange {}.compile(&fxn_input)?); }
-                        _ => unreachable!(),
-                    }
-                }
-                #[cfg(feature = "subscript_range")]
-                [Subscript::Range(ix1), Subscript::Range(ix2)] => {
-                    let result = subscript_range(&subs[0], env, p)?;
-                    fxn_input.push(result);
-                    let result = subscript_range(&subs[1], env, p)?;
-                    fxn_input.push(result);
-                    #[cfg(feature = "matrix")]
-                    plan.borrow_mut()
-                        .push(MatrixAccessRangeRange {}.compile(&fxn_input)?);
-                }
-                #[cfg(all(feature = "subscript_range", feature = "subscript_formula"))]
-                [Subscript::All, Subscript::Formula(ix2)] => {
-                    fxn_input.push(Value::IndexAll);
-                    let result = subscript_formula_ix(&subs[1], env, p)?;
-                    let shape = result.shape();
-                    fxn_input.push(result);
-                    match &shape[..] {
-                        #[cfg(feature = "matrix")]
-                        [1, 1] => { plan.borrow_mut().push(MatrixAccessAllScalar {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "matrix")]
-                        [1, n] => { plan.borrow_mut().push(MatrixAccessAllRange {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "matrix")]
-                        [n, 1] => { plan.borrow_mut().push(MatrixAccessAllRange {}.compile(&fxn_input)?); }
-                        _ => todo!(),
-                    }
-                }
-                #[cfg(all(feature = "subscript_range", feature = "subscript_formula"))]
-                [Subscript::Formula(ix1), Subscript::All] => {
-                    let result = subscript_formula_ix(&subs[0], env, p)?;
-                    let shape = result.shape();
-                    fxn_input.push(result);
-                    fxn_input.push(Value::IndexAll);
-                    match &shape[..] {
-                        #[cfg(feature = "matrix")]
-                        [1, 1] => { plan.borrow_mut().push(MatrixAccessScalarAll {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "matrix")]
-                        [1, n] => { plan.borrow_mut().push(MatrixAccessRangeAll {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "matrix")]
-                        [n, 1] => { plan.borrow_mut().push(MatrixAccessRangeAll {}.compile(&fxn_input)?); }
-                        _ => todo!(),
-                    }
-                }
-                #[cfg(all(feature = "subscript_range", feature = "subscript_formula"))]
-                [Subscript::Range(ix1), Subscript::Formula(ix2)] => {
-                    let result = subscript_range(&subs[0], env, p)?;
-                    fxn_input.push(result);
-                    let result = subscript_formula_ix(&subs[1], env, p)?;
-                    let shape = result.shape();
-                    fxn_input.push(result);
-                    match &shape[..] {
-                        #[cfg(feature = "matrix")]
-                        [1, 1] => { plan.borrow_mut().push(MatrixAccessRangeScalar {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "matrix")]
-                        [1, n] => { plan.borrow_mut().push(MatrixAccessRangeRange {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "matrix")]
-                        [n, 1] => { plan.borrow_mut().push(MatrixAccessRangeRange {}.compile(&fxn_input)?); }
-                        _ => todo!(),
-                    }
-                }
-                #[cfg(all(feature = "subscript_range", feature = "subscript_formula"))]
-                [Subscript::Formula(ix1), Subscript::Range(ix2)] => {
-                    let result = subscript_formula_ix(&subs[0], env, p)?;
-                    let shape = result.shape();
-                    fxn_input.push(result);
-                    let result = subscript_range(&subs[1], env, p)?;
-                    fxn_input.push(result);
-                    match &shape[..] {
-                        #[cfg(feature = "matrix")]
-                        [1, 1] => { plan.borrow_mut().push(MatrixAccessScalarRange {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "matrix")]
-                        [1, n] => { plan.borrow_mut().push(MatrixAccessRangeRange {}.compile(&fxn_input)?); }
-                        #[cfg(feature = "matrix")]
-                        [n, 1] => { plan.borrow_mut().push(MatrixAccessRangeRange {}.compile(&fxn_input)?); }
-                        _ => todo!(),
-                    }
-                }
-                #[cfg(feature = "subscript_range")]
-                [Subscript::All, Subscript::Range(ix2)] => {
-                    fxn_input.push(Value::IndexAll);
-                    let result = subscript_range(&subs[1], env, p)?;
-                    fxn_input.push(result);
-                    #[cfg(feature = "matrix")]
-                    plan.borrow_mut()
-                        .push(MatrixAccessAllRange {}.compile(&fxn_input)?);
-                }
-                #[cfg(feature = "subscript_range")]
-                [Subscript::Range(ix1), Subscript::All] => {
-                    let result = subscript_range(&subs[0], env, p)?;
-                    fxn_input.push(result);
-                    fxn_input.push(Value::IndexAll);
-                    #[cfg(feature = "matrix")]
-                    plan.borrow_mut()
-                        .push(MatrixAccessRangeAll {}.compile(&fxn_input)?);
-                }
-                _ => unreachable!(),
-            };
-            let plan_brrw = plan.borrow();
-            let mut new_fxn = &plan_brrw.last().unwrap();
-            if !expression_solves_deferred(p) {
-              new_fxn.solve();
-            }
-            let res = new_fxn.out();
-            return Ok(res);
-        }
-        _ => unreachable!(),
     }
 }
